cloud function/requirements.txt
requests==2.27.1
jwt==1.3.1
google-auth==2.6.0
google-cloud-secret-manager==2.10.0
domaintools_api==1.0.1
google-api-python-client==2.97.0
google-cloud-storage==2.10.0
redis==5.0.0